backend/requirements.txt
fastapi
uvicorn
pydantic
httpx
python-dotenv
python-multipart
# Add other requirements as needed